+++ src/main/java/com/myecommerce/MyECommerce/exception/ProductException.java
@@ -1,0 +1,14 @@
+package com.myecommerce.MyECommerce.exception;
+
+import com.myecommerce.MyECommerce.exception.errorcode.ProductErrorCode;
+
+public class ProductException extends BaseAbstractException {
+
+    public ProductException(ProductErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public ProductException(ProductErrorCode errorCode, Object... messageArgs) {
+        super(errorCode, messageArgs);
+    }
+}
