+++ src/main/java/com/myecommerce/MyECommerce/service/cart/CartService.java
@@ -7,7 +7,7 @@ import com.myecommerce.MyECommerce.dto.cart.ServiceCartDto;
 import com.myecommerce.MyECommerce.entity.member.Member;
 import com.myecommerce.MyECommerce.entity.product.ProductOption;
 import com.myecommerce.MyECommerce.exception.CartException;
-import com.myecommerce.MyECommerce.exception.ProductionException;
+import com.myecommerce.MyECommerce.exception.ProductException;
 import com.myecommerce.MyECommerce.mapper.ServiceCartMapper;
 import com.myecommerce.MyECommerce.repository.product.ProductOptionRepository;
 import com.myecommerce.MyECommerce.service.redis.RedisMultiDataService;
@@ -20,7 +20,7 @@ import java.time.LocalDate;
 import java.util.Map;
 
 import static com.myecommerce.MyECommerce.exception.errorcode.CartErrorCode.CART_SIZE_EXCEEDED;
-import static com.myecommerce.MyECommerce.exception.errorcode.ProductionErrorCode.PRODUCT_NOT_EXIST;
+import static com.myecommerce.MyECommerce.exception.errorcode.ProductErrorCode.PRODUCT_NOT_EXIST;
 import static com.myecommerce.MyECommerce.type.RedisNamespaceType.CART;
 
 @Service
@@ -108,7 +108,7 @@ public class CartService {
         // 상품옵션조회
         ProductOption foundOption =
                 productOptionRepository.findById(optionId)
-                        .orElseThrow(() -> new ProductionException(PRODUCT_NOT_EXIST));
+                        .orElseThrow(() -> new ProductException(PRODUCT_NOT_EXIST));
 
         // entity -> dto 변환
         return serviceCartMapper.toDto(foundOption);
