+++ src/main/java/com/myecommerce/MyECommerce/service/product/ProductPolicy.java
@@ -4,7 +4,7 @@ import com.myecommerce.MyECommerce.dto.production.ServiceProductionDto;
 import com.myecommerce.MyECommerce.dto.production.ServiceProductionOptionDto;
 import com.myecommerce.MyECommerce.entity.member.Member;
 import com.myecommerce.MyECommerce.entity.product.Product;
-import com.myecommerce.MyECommerce.exception.ProductionException;
+import com.myecommerce.MyECommerce.exception.ProductException;
 import com.myecommerce.MyECommerce.repository.product.ProductOptionRepository;
 import com.myecommerce.MyECommerce.repository.product.ProductRepository;
 import com.myecommerce.MyECommerce.type.ProductionSaleStatusType;
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.myecommerce.MyECommerce.exception.errorcode.ProductionErrorCode.*;
+import static com.myecommerce.MyECommerce.exception.errorcode.ProductErrorCode.*;
 import static com.myecommerce.MyECommerce.type.ProductionSaleStatusType.DELETION;
 
 @Component
@@ -51,7 +51,7 @@ public class ProductPolicy {
                                                     String productCode) {
         productRepository.findBySellerAndCode(sellerId, productCode)
                 .ifPresent(existingProduction -> {
-                    throw new ProductionException(PRODUCT_CODE_ALREADY_REGISTERED);
+                    throw new ProductException(PRODUCT_CODE_ALREADY_REGISTERED);
                 });
     }
 
@@ -65,7 +65,7 @@ public class ProductPolicy {
 
         // 1. 입력받은 옵션코드목록 중 중복 옵션코드 체크
         if (deduplicatedOptionCodes.size() != optionDtoList.size()) {
-            throw new ProductionException(PRODUCT_OPTION_CODE_DUPLICATED);
+            throw new ProductException(PRODUCT_OPTION_CODE_DUPLICATED);
         }
 
         // 2. 이미 등록된 옵션코드와 중복 체크
@@ -73,14 +73,14 @@ public class ProductPolicy {
                 productOptionRepository.findByProductCodeAndOptionCodeIn(
                         productCode, deduplicatedOptionCodes.stream().toList());
         if (!duplicatedOptions.isEmpty()) {
-            throw new ProductionException(PRODUCT_OPTION_CODE_ALREADY_REGISTERED);
+            throw new ProductException(PRODUCT_OPTION_CODE_ALREADY_REGISTERED);
         }
     }
 
     // 상품 수정 시 판매상태 정책
     private void enforceProductModifySaleStatusPolicy(ProductionSaleStatusType saleStatus) {
         if (saleStatus == DELETION) {
-            throw new ProductionException(PRODUCT_ALREADY_DELETED);
+            throw new ProductException(PRODUCT_ALREADY_DELETED);
         }
     }
